requirements_dev.txt
-r requirements.txt

Sphinx==2.3.1
black==19.10b0
codecov==2.0.15
coverage==5.0.3
flake8==3.7.9
grip==4.5.2
isort==4.3.21
mypy==0.761
pytest-aiohttp==0.3.0
pytest-cov==2.8.1
pytest-xprocess==0.13.1
pytest==5.3.5
sphinx-autodoc-typehints==1.10.3

setuptools
twine==3.1.1
wheel
